Modified template for when displayed on subordinate_tickets.php

+++ src/includes/alerts_template.php
@@ -5,11 +5,22 @@ function display_ticket_alerts($alerts)
     foreach ($alerts as $alert) {
 ?>
         <p class="<?= $alert["alert_level"] ?>">
-            <a href="/controllers/tickets/edit_ticket.php?id=<?= $alert["ticket_id"] ?>">Ticket: <?= $alert["ticket_id"].' -' ?>
-            <?= $alert["message"] ?>
+            <a href="/controllers/tickets/edit_ticket.php?id=<?= $alert["ticket_id"] ?>">
+                <?php if (basename($_SERVER['PHP_SELF']) == 'subordinate_tickets.php') {
+                    echo  $alert["employee"] . ' - ';
+                } ?>
+                Ticket: <?= $alert["ticket_id"] . ' -' ?>
+                <?= $alert["message"] ?>
             </a>
-            <!-- <a href="/controllers/tickets/alert_delete.php?id=<?= $alert["id"] ?>"> Delete</a> -->
-            <a class="close-alert" href="/controllers/tickets/alert_delete.php?id=<?= $alert["id"] ?>">&times;</a>
+
+            <?php
+            if (basename($_SERVER['PHP_SELF']) != 'subordinate_tickets.php') {
+            ?>
+                <a class="close-alert" href="/controllers/tickets/alert_delete.php?id=<?= $alert["id"] ?>">&times;</a>
+            <?php
+            }
+            ?>
+
         </p>
 <?php
